refactor(communities): rename entity file and update imports for Community entity
