+++ src/communities/communities.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { CommunitiesService } from './communities.service';
 import { CommunitiesController } from './communities.controller';
-import { Community } from './communities.entity';
+import { Community } from './community.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Community])],
