+++ src/communities/communities.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Community } from './communities.entity';
+import { Community } from './community.entity';
 
 @Injectable()
 export class CommunitiesService {
